* WIP Plugin folder structure (mirroring backend) in frontend code   * DONE refactor: move register_components() outside of main.js
